ParallelGraph: move lac code to a separate module
